modified with rubocop and added new comments

+++ app/workers/clean_images.rb
@@ -1,10 +1,9 @@
+# worker for sidekiq to delete rejected images
 class CleanImages
   include Sidekiq::Worker
 
   def perform(id_image)
     image = Image.find_by(id: id_image)
-    if image.rejected?
-      image.destroy!
-    end
+    image.destroy! if image.rejected?
   end
 end
